+++ kubejs/startup_scripts/0_consts.js
@@ -1,1 +1,2 @@
 global.colors = ["white", "orange", "magenta", "light_blue", "yellow", "lime", "pink", "gray", "light_gray", "cyan", "purple", "blue", "brown", "green", "red", "black"];
+global.regexAny = /^[\s\S]*$/;